remove adatypes
