+++ app/actions/ada/trezor-send-actions.js
@@ -1,6 +1,6 @@
 // @flow
 import Action from '../lib/Action';
-import type { BaseSignRequest } from '../../api/ada/adaTypes';
+import type { BaseSignRequest } from '../../api/ada/transactions/types';
 
 export type SendUsingTrezorParams = {
   signRequest: BaseSignRequest,
