+++ app/api/ada/transactions/byron/transactionV2.test.js
@@ -3,8 +3,9 @@
 import '../../lib/test-config';
 import { schema } from 'lovefield';
 import type {
-  RemoteUnspentOutput, AddressedUtxo,
-} from '../../adaTypes';
+  AddressedUtxo,
+} from '../types';
+import type { RemoteUnspentOutput } from '../../lib/state-fetch/types';
 import {
   newAdaUnsignedTx,
   newAdaUnsignedTxFromUtxo,
