+++ app/components/wallet/send/WalletSendConfirmationDialog.js
@@ -21,7 +21,7 @@ import RawHash from '../../widgets/hashWrappers/RawHash';
 import type { ExplorerType } from '../../../domain/Explorer';
 
 import WarningBox from '../../widgets/WarningBox';
-import type { BaseSignRequest } from '../../../api/ada/adaTypes';
+import type { BaseSignRequest } from '../../../api/ada/transactions/types';
 
 const messages = defineMessages({
   walletPasswordLabel: {
