+++ app/stores/ada/AdaTransactionBuilderStore.js
@@ -8,8 +8,8 @@ import type {
   CreateUnsignedTxFunc,
 } from '../../api/ada';
 
-import type { BaseSignRequest } from '../../api/ada/adaTypes';
-import { signRequestFee, signRequestTotalInput } from '../../api/ada/lib/utils';
+import type { BaseSignRequest } from '../../api/ada/transactions/types';
+import { signRequestFee, signRequestTotalInput } from '../../api/ada/transactions/utils';
 import {
   asGetAllUtxos, asHasChains,
 } from '../../api/ada/lib/storage/models/PublicDeriver/index';
